fix: redact anchor-failure diagnostics and fail closed on empty paths
Set RedactPaths on pre-SubmitReview anchor authority failures and reject
inline payloads whose paths all trim empty. Split oversized contract tests
under the AGENTS.md 300-line test-file threshold.

Generated-By: looper 0.10.5 (runner=fixer, agent=grok-build)
